adding LoopStatRecordStorage
