+++ FreeAPS/Sources/APS/APSManager.swift
@@ -76,6 +76,7 @@ actor BaseAPSManager: APSManager, LifetimeOwner, AppService {
     private let overrideStorage: OverrideStorage
     private let overrideManager: OverrideManager
     private let loopEventsStorage: LoopEventsStorage
+    private let loopStatRecordStorage: LoopStatRecordStorage
 
     @Persisted(key: "lastAutotuneDate") private var lastAutotuneDate = Date()
     @Persisted(key: "lastStartLoopDate") private var lastStartLoopDate: Date = .distantPast
@@ -109,7 +110,8 @@ actor BaseAPSManager: APSManager, LifetimeOwner, AppService {
         overrideStorage: OverrideStorage,
         overrideManager: OverrideManager,
         dynamicStateManager: DynamicStateManager,
-        loopEventsStorage: LoopEventsStorage
+        loopEventsStorage: LoopEventsStorage,
+        loopStatRecordStorage: LoopStatRecordStorage,
     ) {
         self.appCoordinator = appCoordinator
         self.storage = storage
@@ -127,6 +129,7 @@ actor BaseAPSManager: APSManager, LifetimeOwner, AppService {
         self.overrideManager = overrideManager
         self.dynamicStateManager = dynamicStateManager
         self.loopEventsStorage = loopEventsStorage
+        self.loopStatRecordStorage = loopStatRecordStorage
     }
 
     // this is called at the app start
@@ -265,7 +268,7 @@ actor BaseAPSManager: APSManager, LifetimeOwner, AppService {
                 loopStatRecord.loopStatus = error
             }
 
-            await self.persistLoopStats(loopStatRecord: loopStatRecord, error: loopOutcome.error)
+            await self.loopStatRecordStorage.persistLoopStats(loopStatRecord: loopStatRecord, error: loopOutcome.error)
 
             await self.updateIOB(pumpHistory: appCoordinator.pumpHistory.value)
 
@@ -826,21 +829,6 @@ actor BaseAPSManager: APSManager, LifetimeOwner, AppService {
     private func activeBolusView() -> Bool {
         let defaults = UserDefaults.standard
         return defaults.bool(forKey: IAPSconfig.inBolusView)
-    }
-
-    private func persistLoopStats(loopStatRecord: LoopStats, error: String?) async {
-        await CoreDataStack.shared.persistentContainer.performBackgroundTask { coredataContext in
-            let nLS = LoopStatRecord(context: coredataContext)
-            nLS.start = loopStatRecord.start
-            nLS.end = loopStatRecord.end ?? Date()
-            nLS.loopStatus = loopStatRecord.loopStatus
-            nLS.duration = loopStatRecord.duration ?? 0.0
-            nLS.interval = loopStatRecord.interval ?? 0.0
-            if let error = error {
-                nLS.error = error
-            }
-            try? coredataContext.save()
-        }
     }
 
     private func processError(_ error: Error) {
