+++ FreeAPS/Sources/Assemblies/ServiceAssembly.swift
@@ -112,6 +112,7 @@ final class ServiceAssembly: Assembly {
             )
         }
         container.register(CoreDataManager.self) { r in CoreDataManager(resolver: r) }
+        container.register(LoopStatRecordStorage.self) { r in LoopStatRecordStorage(resolver: r) }
 
         container.register(AppServices.self) { r in
             nonisolated(unsafe) let resolver = r
